zephyr: lib: Remove ctx_alloc.h and put the contents into rtos/alloc.h
Remove the ctx_alloc.h header and move the implementation into
rtos/alloc.h.
